successfully refactored human cases
