+++ src/components/cases/NewCaseModal.jsx
@@ -1,0 +1,401 @@
+import React from "react";
+import { X, Save } from "lucide-react";
+
+const NewCaseModal = ({
+  isOpen,
+  onClose,
+  formData,
+  onInputChange,
+  onSubmit,
+  isSubmitting,
+}) => {
+  if (!isOpen) return null;
+
+  return (
+    <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50 p-4">
+      <div className="bg-white rounded-xl shadow-xl max-w-4xl w-full max-h-[90vh] overflow-y-auto">
+        {/* Modal Header */}
+        <div className="flex items-center justify-between p-6 border-b border-gray-200">
+          <h2 className="text-xl font-semibold text-gray-900">
+            Create New Case
+          </h2>
+          <button
+            onClick={onClose}
+            className="text-gray-400 hover:text-gray-600 transition-colors"
+          >
+            <X className="w-6 h-6" />
+          </button>
+        </div>
+
+        {/* Modal Body */}
+        <form onSubmit={onSubmit} className="p-6">
+          <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
+            {/* Case Information */}
+            <div className="md:col-span-3">
+              <h3 className="text-lg font-medium text-gray-900 mb-4">
+                Case Information
+              </h3>
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-2">
+                Case ID
+              </label>
+              <input
+                type="text"
+                name="case_id"
+                value={formData.case_id}
+                onChange={onInputChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+                required
+              />
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-2">
+                Person ID
+              </label>
+              <input
+                type="text"
+                name="personID"
+                value={formData.personID}
+                onChange={onInputChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+                required
+              />
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-2">
+                Disease
+              </label>
+              <select
+                name="disease"
+                value={formData.disease}
+                onChange={onInputChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+                required
+              >
+                <option value="">Select Disease</option>
+                <option value="Yellow Fever">Yellow Fever</option>
+                <option value="Lassa Fever">Lassa Fever</option>
+                <option value="Ebola">Ebola</option>
+                <option value="Monkey Pox">Monkey Pox</option>
+                <option value="Cholera">Cholera</option>
+                <option value="Meningitis">Meningitis</option>
+                <option value="COVID-19">COVID-19</option>
+              </select>
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-2">
+                Classification
+              </label>
+              <select
+                name="classification"
+                value={formData.classification}
+                onChange={onInputChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+                required
+              >
+                <option value="">Select Classification</option>
+                <option value="Suspect">Suspect</option>
+                <option value="Probable">Probable</option>
+                <option value="Confirmed">Confirmed</option>
+              </select>
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-2">
+                Outcome
+              </label>
+              <select
+                name="outcome"
+                value={formData.outcome}
+                onChange={onInputChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+              >
+                <option value="">Select Outcome</option>
+                <option value="Recovered">Recovered</option>
+                <option value="Deceased">Deceased</option>
+                <option value="Under Treatment">Under Treatment</option>
+                <option value="No Outcome Yet">No Outcome Yet</option>
+              </select>
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-2">
+                Category
+              </label>
+              <select
+                name="category"
+                value={formData.category}
+                onChange={onInputChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+              >
+                <option value="Human">Human</option>
+                <option value="Animal">Animal</option>
+                <option value="Environmental">Environmental</option>
+              </select>
+            </div>
+
+            {/* Location Information */}
+            <div className="md:col-span-3">
+              <h3 className="text-lg font-medium text-gray-900 mb-4 mt-6">
+                Location Information
+              </h3>
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-2">
+                State
+              </label>
+              <input
+                type="text"
+                name="state"
+                value={formData.state}
+                onChange={onInputChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+                required
+              />
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-2">
+                LGA
+              </label>
+              <input
+                type="text"
+                name="lga"
+                value={formData.lga}
+                onChange={onInputChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+                required
+              />
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-2">
+                Health Facility
+              </label>
+              <input
+                type="text"
+                name="health_facility"
+                value={formData.health_facility}
+                onChange={onInputChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+                required
+              />
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-2">
+                Region
+              </label>
+              <input
+                type="text"
+                name="region"
+                value={formData.region}
+                onChange={onInputChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+                required
+              />
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-2">
+                Longitude
+              </label>
+              <input
+                type="number"
+                step="any"
+                name="longitude"
+                value={formData.longitude}
+                onChange={onInputChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+              />
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-2">
+                Latitude
+              </label>
+              <input
+                type="number"
+                step="any"
+                name="latitude"
+                value={formData.latitude}
+                onChange={onInputChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+              />
+            </div>
+
+            {/* Patient Information */}
+            <div className="md:col-span-3">
+              <h3 className="text-lg font-medium text-gray-900 mb-4 mt-6">
+                Patient Information
+              </h3>
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-2">
+                Age
+              </label>
+              <input
+                type="number"
+                name="age"
+                value={formData.age}
+                onChange={onInputChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+                required
+              />
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-2">
+                Sex
+              </label>
+              <select
+                name="sex"
+                value={formData.sex}
+                onChange={onInputChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+                required
+              >
+                <option value="">Select Sex</option>
+                <option value="Male">Male</option>
+                <option value="Female">Female</option>
+              </select>
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-2">
+                Occupation
+              </label>
+              <input
+                type="text"
+                name="occupation"
+                value={formData.occupation}
+                onChange={onInputChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+              />
+            </div>
+
+            {/* Clinical Information */}
+            <div className="md:col-span-3">
+              <h3 className="text-lg font-medium text-gray-900 mb-4 mt-6">
+                Clinical Information
+              </h3>
+            </div>
+
+            <div className="md:col-span-2">
+              <label className="block text-sm font-medium text-gray-700 mb-2">
+                Symptoms
+              </label>
+              <textarea
+                name="symptoms"
+                value={formData.symptoms}
+                onChange={onInputChange}
+                rows={3}
+                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+                placeholder="Describe symptoms separated by commas"
+              />
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-2">
+                Risk Factors
+              </label>
+              <textarea
+                name="risk_factors"
+                value={formData.risk_factors}
+                onChange={onInputChange}
+                rows={3}
+                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+                placeholder="Describe risk factors"
+              />
+            </div>
+
+            {/* Dates and Reporting */}
+            <div className="md:col-span-3">
+              <h3 className="text-lg font-medium text-gray-900 mb-4 mt-6">
+                Dates and Reporting
+              </h3>
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-2">
+                Date of Onset
+              </label>
+              <input
+                type="datetime-local"
+                name="date_of_onset"
+                value={formData.date_of_onset}
+                onChange={onInputChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+                required
+              />
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-2">
+                Date of Confirmation
+              </label>
+              <input
+                type="datetime-local"
+                name="date_of_confirmation"
+                value={formData.date_of_confirmation}
+                onChange={onInputChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+              />
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-2">
+                Reporting Source
+              </label>
+              <input
+                type="text"
+                name="reporting_source"
+                value={formData.reporting_source}
+                onChange={onInputChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+                placeholder="e.g., Hospital Report, Community Health Worker"
+              />
+            </div>
+          </div>
+
+          {/* Modal Footer */}
+          <div className="flex items-center justify-end gap-4 mt-8 pt-6 border-t border-gray-200">
+            <button
+              type="button"
+              onClick={onClose}
+              className="px-4 py-2 text-gray-700 bg-gray-100 rounded-lg hover:bg-gray-200 transition-colors"
+            >
+              Cancel
+            </button>
+            <button
+              type="submit"
+              disabled={isSubmitting}
+              className="flex items-center gap-2 px-6 py-2 bg-blue-600 text-white rounded-lg hover:bg-blue-700 transition-colors disabled:opacity-50 disabled:cursor-not-allowed"
+            >
+              {isSubmitting ? (
+                <>
+                  <div className="w-4 h-4 border-2 border-white border-t-transparent rounded-full animate-spin"></div>
+                  Creating...
+                </>
+              ) : (
+                <>
+                  <Save className="w-4 h-4" />
+                  Create Case
+                </>
+              )}
+            </button>
+          </div>
+        </form>
+      </div>
+    </div>
+  );
+};
+
+export default NewCaseModal;
